requirements.txt
openai
python-dotenv
azure-identity
pydantic
azure-ai-projects
azure-ai-inference
rich